+++ data/src/main/java/com/gdg/data/mapper/ResponseWeatherDtoMapper.kt
@@ -4,5 +4,5 @@ import com.gdg.data.dto.response.WeatherResponseDto
 import com.gdg.domain.entity.WeatherEntity
 
 fun WeatherResponseDto.toWeatherEntity() = WeatherEntity(
-    id, areaNm, skyStts, temp, pm25Index, pm10Index
+    id.toInt(), areaNm, skyStts, temp, pm25Index, pm10Index
 )
